requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# PDF Processing
PyMuPDF>=1.23.0
Pillow>=10.0.0
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0

# Security (optional - for enhanced file type validation)
# On Windows: pip install python-magic-bin
# On Linux: pip install python-magic (requires: apt-get install libmagic1)
# On macOS: pip install python-magic (requires: brew install libmagic)
# python-magic-bin>=0.4.14  # Uncomment for Windows
